% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_weights.R
\name{calculate_log_weights}
\alias{calculate_log_weights}
\title{Importance weight calculation}
\usage{
calculate_log_weights(model, fit, old_data, new_data)
}
\arguments{
\item{model}{Path to .stan file containing model}

\item{fit}{`CmdStanMCMC` fit object or array of posterior samples}

\item{old_data}{Corresponding data list used to generate `fit`}

\item{new_data}{Data list for which to calculate importance weights}
}
\value{
log-importance weights for the posterior samples
}
\description{
Calculate log importance weights
}
\examples{
logweights <- calculate_log_weights(model, fit, old_data, new_data)
}
