% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resample}
\alias{resample}
\title{Resampling}
\usage{
resample(fit, log_weights)
}
\arguments{
\item{fit}{`CmdStanMCMC` fit object or array of posterior samples}

\item{log_weights}{Log-importance weights}
}
\value{
Resampled posterior samples
}
\description{
Resample posterior samples using log-importance weights
}
\examples{
resampled_samples <- resample(fit, log_weights)
}
